Mark cmake Multi-Arch: foreign.
The change was stolen from Ubuntu cmake_2.8.7-0ubuntu3 package. Thanks to Steve
Langasek who originally wrote:

Mark cmake Multi-Arch: foreign; it exports a cross-build-aware interface
(with a little bit of cross-configuration from dpkg-cross) so we should
allow the native version to satisfy our dependencies, instead of trying
to pull in the host-architecture version when cross-building.